Added check for last modified time to settings CLI (#3403)
* Added check for last modified time to settings CLI

+++ prime-router/build.gradle.kts
@@ -110,8 +110,9 @@ tasks.clean {
     delete("target")
     // clean up all the old event files in the SOAP set up
     doLast {
-        val eventsDir = File("../.environment/soap_service/soap/event/v1/")
-        if (eventsDir.exists()) {
+        val eventsDir = File("../.environment/soap_service/soap/event/v1")
+        if (eventsDir.exists() && eventsDir.isDirectory && eventsDir.listFiles().isNotEmpty()) {
+            // Note FileUtils does not like when the folder is empty.
             FileUtils.listFiles(eventsDir, arrayOf("event"), true).forEach {
                 it.delete()
             }
